refactor(api/v1/controllers/quizplayercontroller.js): complete `ctAddQuizPlayer` function
completed this end point which allows a player to add themselves to a valid quiz; get details about
the quiz in a response if the request is successful

BREAKING CHANGE: user currently gets the quiz questions in a response even if the quiz occurs in the
future. This needs to be removed so that only quizzes that are currently running provide the user
with a list of questions in this response object.

re #47